backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.7.0

# SQLAlchemy async
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0

# Database drivers
aiosqlite>=0.20.0       # SQLite dev
asyncpg>=0.30.0         # PostgreSQL / TimescaleDB prod

# Redis
redis[asyncio]>=5.2.0
hiredis>=3.0.0

# HTTP
httpx>=0.28.0
python-multipart>=0.0.12

# Utils
tzdata>=2024.1               # Windows timezone database (zoneinfo)
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
